refactor: organize file structure for TopPage
