+++ client/components/page/Top/Top.page.tsx
@@ -1,0 +1,8 @@
+import { TopPageProps } from 'pages';
+import Top from './Top';
+
+const TopPage = (props: TopPageProps) => {
+  return <Top {...props} />;
+};
+
+export default TopPage;
