+++ client/components/page/Top/index.tsx
@@ -1,0 +1,1 @@
+export { default as TopPage } from './Top.page';
