+++ client/components/shared/footer/SidebarWithHeader.tsx
